feat: move cli design tokens to paragon cli
