+++ lib/build-scss.js
@@ -1,4 +1,3 @@
-#!/usr/bin/env node
 const fs = require('fs');
 const sass = require('sass');
 const postCSS = require('postcss');
@@ -8,7 +7,7 @@ const postCSSMinify = require('postcss-minify');
 const combineSelectors = require('postcss-combine-duplicated-selectors');
 const { pathToFileURL } = require('url');
 const path = require('path');
-const { program, Option } = require('commander');
+const minimist = require('minimist');
 
 const paragonThemeOutputFilename = 'theme-urls.json';
 
@@ -133,79 +132,32 @@ const compileAndWriteStyleSheets = ({
     });
 };
 
-program
-  .version('0.0.1')
-  .description('CLI to compile Paragon\'s core and themes\' SCSS into CSS.')
-  .addOption(
-    new Option(
-      '--corePath <corePath>',
-      'Path to the theme\'s core SCSS file, defaults to Paragon\'s core.scss.',
-    ),
-  )
-  .addOption(
-    new Option(
-      '--themesPath <themesPath>',
-      `Path to the directory that contains themes' files. Expects directory to have following structure:
-      themes/ 
-        light/
-        │  ├─ index.css
-        │  ├─ other_css_files
-        dark/
-        │  ├─ index.css
-        │  ├─ other_css_files
-        some_other_custom_theme/
-        │  ├─ index.css
-        │  ├─ other_css_files
-        ...
-      
-      where index.css has imported all other CSS files in the theme's subdirectory. The script will output
-      light.css, dark.css and some_other_custom_theme.css files (together with maps and minified versions).
-      You can provide any amount of themes. Default to paragon's themes.
-      `,
-    ),
-  )
-  .addOption(
-    new Option(
-      '--outDir <outDir>',
-      'Specifies directory where to out resulting CSS files.',
-    ),
-  )
-  .addOption(
-    new Option(
-      '--defaultThemeVariants <defaultThemeVariants...>',
-      `Specifies default theme variants. Defaults to a single 'light' theme variant.
-      You can provide multiple default theme variants by passing multiple values, for
-      example: \`--defaultThemeVariants light dark\`
-      `,
-    ),
-  );
+function buildScssCommand(commandArgs) {
+  const argv = minimist(commandArgs);
+  const corePath = argv.corePath || path.resolve(process.cwd(), 'styles/scss/core/core.scss');
+  const themesPath = argv.themesPath || path.resolve(process.cwd(), 'styles/css/themes');
+  const outDir = argv.outDir || './dist';
+  const defaultThemeVariants = argv.defaultThemeVariants ? argv.defaultThemeVariants.split(',') : ['light'];
 
-program.parse(process.argv);
-
-const options = program.opts();
-const {
-  corePath = path.resolve(__dirname, 'styles/scss/core/core.scss'),
-  themesPath = path.resolve(__dirname, 'styles/css/themes'),
-  outDir = './dist',
-  defaultThemeVariants = ['light'],
-} = options;
-
-// Core CSS
-compileAndWriteStyleSheets({
-  name: 'core',
-  stylesPath: corePath,
-  outDir,
-});
-
-// Theme Variants CSS
-fs.readdirSync(themesPath, { withFileTypes: true })
-  .filter((item) => item.isDirectory())
-  .forEach((themeDir) => {
-    compileAndWriteStyleSheets({
-      name: themeDir.name,
-      stylesPath: `${themesPath}/${themeDir.name}/index.css`,
-      outDir,
-      isThemeVariant: true,
-      isDefaultThemeVariant: defaultThemeVariants.includes(themeDir.name),
-    });
+  // Core CSS
+  compileAndWriteStyleSheets({
+    name: 'core',
+    stylesPath: corePath,
+    outDir,
   });
+
+  // Theme Variants CSS
+  fs.readdirSync(themesPath, { withFileTypes: true })
+    .filter((item) => item.isDirectory())
+    .forEach((themeDir) => {
+      compileAndWriteStyleSheets({
+        name: themeDir.name,
+        stylesPath: `${themesPath}/${themeDir.name}/index.css`,
+        outDir,
+        isThemeVariant: true,
+        isDefaultThemeVariant: defaultThemeVariants.includes(themeDir.name),
+      });
+    });
+}
+
+module.exports = buildScssCommand;
