+++ package.json
@@ -6,9 +6,6 @@
   "module": "dist/index.js",
   "license": "Apache-2.0",
   "bin": {
-    "build-design-tokens": "./tokens/build-tokens.js",
-    "replace-scss-with-css": "./tokens/replace-variables.js",
-    "build-scss": "./build-scss.js",
     "paragon": "./bin/paragon-scripts.js"
   },
   "publishConfig": {
@@ -51,9 +48,9 @@
     "build-types": "tsc --emitDeclarationOnly",
     "playroom:start": "npm run playroom:start --workspace=www",
     "playroom:build": "npm run playroom:build --workspace=www",
-    "build-tokens": "node tokens/build-tokens.js --build-dir ./styles/css",
-    "replace-variables-usage-with-css": "node tokens/replace-variables.js -p src -t usage",
-    "replace-variables-definition-with-css": "node tokens/replace-variables.js -p src -t definition"
+    "build-tokens": "./bin/paragon-scripts.js build-tokens --build-dir ./styles/css",
+    "replace-variables-usage-with-css": "./bin/paragon-scripts.js replace-variables -p src -t usage",
+    "replace-variables-definition-with-css": "./bin/paragon-scripts.js replace-variables -p src -t definition"
   },
   "dependencies": {
     "@popperjs/core": "^2.11.4",
@@ -69,6 +66,7 @@
     "inquirer": "^8.2.5",
     "lodash.uniqby": "^4.7.0",
     "mailto-link": "^2.0.0",
+    "minimist": "^1.2.8",
     "postcss": "^8.4.21",
     "postcss-combine-duplicated-selectors": "^10.0.3",
     "postcss-custom-media": "^9.1.2",
